+++ ui/components/boards/metrics.mjs
@@ -1,5 +1,5 @@
 // Dependencies
-import { formatBytes, timeAgo, parseJson } from 'lib/utils.mjs'
+import { cloneAsPojo, formatBytes, timeAgo, parseJson } from 'lib/utils.mjs'
 import orderBy from 'lodash/orderBy.js'
 import { chartTemplates } from './chart-templates.mjs'
 import { linkClasses } from 'components/link.mjs'
@@ -18,6 +18,7 @@ import { KeyVal } from 'components/keyval.mjs'
 import { ToggleLiveButton } from 'components/boards/shared.mjs'
 import { ChartsProvider } from './charts-provider.mjs'
 import { Echart } from 'components/echarts.mjs'
+import { Popout } from 'components/popout.mjs'
 
 /**
  * This compnent renders a table with the host for which we have cached metrics
@@ -195,7 +196,7 @@ export const HostMetricsTable = ({ host, module = false }) => {
         <thead>
           <tr>
             {cols.map((field) => (
-              <th key={field} className="text-left" key={field}>
+              <th key={field} className="text-left">
                 <button
                   className={`btn btn-link capitalize px-0 ${linkClasses}`}
                   onClick={() => (order === field ? setDesc(!desc) : setOrder(field))}
@@ -273,8 +274,9 @@ const transformMetrics = ({ host, module, metricset, data, templates }) =>
         data,
         templates,
         clone,
+        formatBytes,
       })
-    : data
+    : { err: 'noTransformAvailable', data }
 
 /**
  * This component renders visualisations for all cached
@@ -298,18 +300,18 @@ const ShowMetricsInner = ({ host, module, metricset, hostname, show }) => {
     refetchIntervalInBackground: false,
   })
 
-  // Defer to chart transformer
-  const data = parseCachedMetrics(cache)
-
   // Don't bother if there's nothing in the cache
-  return !cache || cache.length < 1 ? (
+  if (!cache || cache.length < 1) return (
     <>
       <Loading />
       <p>Nothing in the cache to show you here.</p>
     </>
-  ) : (
-    <EchartWrapper {...{ data, host, module, metricset, paused, setPaused, hostname, show }} />
   )
+
+  // Defer to chart transformer
+  const data = parseCachedMetrics(cache)
+
+  return <EchartWrapper {...{ data, host, module, metricset, paused, setPaused, hostname, show }} />
 }
 
 async function runShowMetricsApiCall(api, host, module, metricset) {
@@ -342,7 +344,7 @@ const EchartWrapper = ({
         module,
         metricset,
         data,
-        templates: chartTemplates(data),
+        templates: cloneAsPojo(chartTemplates),
       }),
     [host, module, metricset, data]
   )
@@ -357,6 +359,15 @@ const EchartWrapper = ({
 
   const isEnabled = (option, i) =>
     enabled === true || (enabled && (enabled[i] || enabled[option?.id])) ? true : false
+
+  if (option.err === 'noTransformAvailable') return (
+    <Popout note>
+      <h5>No visualisations available</h5>
+      <p>No charts are loaded for the <code>{metricset}</code> metricset of the <code>{module}</code> module.</p>
+      <p>If this modules provides chart templates, you may need to preseed them.</p>
+    </Popout>
+  )
+
 
   return (
     <div className="flex flex-col gap-4">
@@ -420,6 +431,7 @@ export const SingleEchart = ({ option, href = false }) => {
  * @return {object} data - The same data parsed
  */
 export function parseCachedMetrics(data) {
+  if (!data) return data
   if (Array.isArray(data))
     return orderBy(
       data.map((entry) => parseJson(entry)),
@@ -427,6 +439,6 @@ export function parseCachedMetrics(data) {
       'ASC'
     )
 
-  console.log('Metrics data was a not an array. This is unexpected')
+  console.log('Metrics data was a not an array. This is unexpected', data)
   return []
 }